# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 自炊支援・食費節約アプリ API
  description: |
    一人暮らしまたは2〜4人家族の世帯を対象とした自炊支援・食費節約Webアプリケーションのバックエンド API仕様書。
    
    ## 主要機能
    - ユーザー管理（登録、ログイン、プロフィール管理）
    - レシピ管理（作成、検索、編集、削除）
    - レビュー機能（星評価、コメント、通報）
    - スケジュール管理（料理予定と実績）
    - 買い物リスト管理
    - サボり防止アラート
    - 管理者機能
    
    ## 認証
    - AWS Cognito認証を使用
    - アクセストークンの有効期限: 1時間
    - リフレッシュトークンの有効期限: 90日
    - ユーザーIDはJWTトークンのsubクレームから取得（SecurityContext）
    
    ## 多言語対応
    - 日本語（ja）、韓国語（ko）
    - Accept-Languageヘッダーでエラーメッセージの言語を指定
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Private
servers:
  - url: http://localhost:8080
    description: ローカル開発環境
  - url: https://api-dev.example.com
    description: 開発環境
  - url: https://api.example.com
    description: 本番環境
...
paths:
...
  /api/ai/chat:
    post:
      tags: [AI Assistant]
      summary: 新規会話開始
      description: 新しい会話を開始してAIとチャットします。ユーザーIDはJWTトークンから取得されます。
      operationId: startChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '201':
          description: 会話開始成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/ai/conversations/{conversationId}/messages:
    get:
      tags: [AI Assistant]
      summary: メッセージ一覧取得
      description: 会話のメッセージ一覧を取得します。ユーザーIDはJWTトークンから取得されます。
      operationId: getMessages
      parameters:
        - $ref: '#/components/parameters/ConversationIdPath'
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageResponseDto'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [AI Assistant]
      summary: メッセージ送信
      description: 既存の会話にメッセージを送信します。ユーザーIDはJWTトークンから取得されます。
      operationId: sendMessage
      parameters:
        - $ref: '#/components/parameters/ConversationIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: 送信成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponseDto'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/ai/conversations/{conversationId}:
    delete:
      tags: [AI Assistant]
      summary: 会話削除
      description: 会話を削除します。ユーザーIDはJWTトークンから取得されます。
      operationId: deleteConversation
      parameters:
        - $ref: '#/components/parameters/ConversationIdPath'
      responses:
        '204':
          description: 削除成功
        '404':
          $ref: '#/components/responses/NotFound'

  # ========================================
  # Alerts API
  # ========================================
...
components:
...
  parameters:
...
    ConversationIdPath:
      name: conversationId
      in: path
      required: true
      description: 会話ID
      schema:
        type: string
        format: uuid

  # ========================================
  # Responses
  # ========================================
  responses:
    BadRequest:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: カレーの作り方を教えて
        conversationType:
          type: string
          enum: [general, recipe_recommendation, expiry_check]
          default: general
          description: |
            会話タイプ:
            - general: 一般的なチャット
            - recipe_recommendation: レシピ推薦（在庫連携）
            - expiry_check: 賞味期限確認（在庫連携）
    SendMessageRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: もっと簡単なレシピはありますか？
    ChatResponseDto:
      type: object
      properties:
        conversationId:
          type: string
          format: uuid
        response:
          type: string
          description: AIからの応答
...
    MessageResponseDto:
      type: object
      properties:
        messageId:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time

    # ----------------------------------------
    # Error Schema
    # ----------------------------------------
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: バリデーションエラー
        path:
          type: string
          example: /api/users/register

  # ========================================
  # Security Schemes
  # ========================================
  securitySchemes:
    CognitoAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        AWS Cognito JWTトークン認証。
        ログインエンドポイント（/api/users/login）から取得したアクセストークンを使用します。
        ユーザーIDはJWTトークンのsubクレームから自動的に取得されます（SecurityContext）。
        
        例: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`

# ========================================
# Global Security (認証不要のエンドポイントは個別にsecurity: []を指定)
# ========================================
security:
  - CognitoAuth: []